requirements.txt
nltk==3.6
rouge-score==0.1
transformers==4.20.1
torch==1.11.0